dosdet/requirements.txt
numpy>=1.26
pandas>=2.0
pyarrow>=16.0.0
scipy>=1.11
scikit-learn>=1.4
python-dateutil>=2.9
tqdm>=4.66
scapy>=2.5.0
torch>=2.2
pyyaml>=6.0.1
matplotlib>=3.8